Company/requirements.txt
cffi==1.15.1
clr-loader==0.2.6
comtypes==1.2.0
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.0
imageio==2.31.1
importlib-resources==6.0.1
imutils==0.5.4
kiwisolver==1.4.4
lazy_loader==0.3
matplotlib==3.7.2
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
pycparser==2.21
pygrabber==0.1
pyparsing==3.0.9
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
python-dateutil==2.8.2
pythonnet==3.0.1
pywin32==306
QtRangeSlider==0.1.5
scikit-image==0.21.0
scipy==1.11.1
shiboken6==6.5.2
six==1.16.0
tifffile==2023.8.12
zipp==3.16.2
